refactor: Moved the GitHub util classes and functions out of the GitHubCallback page and into their own files
