+++ src/features/github/GitHubUtils.ts
@@ -1,0 +1,13 @@
+import { URIField } from './GitHubClasses';
+
+export function encodeAsURL(messageBodyPrototype: URIField[]) {
+  const uriComponents: string[] = [];
+
+  messageBodyPrototype.forEach(element => {
+    uriComponents.push(
+      [encodeURIComponent(element.name || ''), encodeURIComponent(element.value || '')].join('=')
+    );
+  });
+
+  return uriComponents.join('&');
+}
